sys2-backend/requirements.txt
alembic==1.15.2
annotated-types==0.7.0
anyio==4.9.0
bcrypt==4.0.1
cffi==1.17.1
click==8.1.8
cryptography==44.0.2
dnspython==2.7.0
ecdsa==0.19.1
email_validator==2.2.0
fastapi==0.115.12
greenlet==3.1.1
h11==0.14.0
httptools==0.6.4
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
passlib==1.7.4
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.8.1
pydantic_core==2.33.1
python-dotenv==1.1.0
python-http-client==3.3.7
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
rsa==4.9
sendgrid==6.11.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starkbank-ecdsa==2.2.0
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.1
uvloop==0.21.0
watchfiles==1.0.5
websockets==15.0.1
